requirement.txt
google-cloud-datastore==1.6.0
telepot==12.6